+++ swap/src/asb/recovery/cancel.rs
@@ -30,6 +30,8 @@ pub async fn cancel(
         | AliceState::CancelTimelockExpired { monero_wallet_restore_blockheight, transfer_proof, state3}
         | AliceState::BtcCancelled { monero_wallet_restore_blockheight, transfer_proof, state3 }
         | AliceState::BtcRefunded { monero_wallet_restore_blockheight, transfer_proof,  state3 ,.. }
+        | AliceState::BtcPartiallyRefunded { monero_wallet_restore_blockheight, transfer_proof,  state3 ,.. }
+        | AliceState::XmrRefundable { monero_wallet_restore_blockheight, transfer_proof,  state3 ,.. }
         | AliceState::BtcPunishable { monero_wallet_restore_blockheight, transfer_proof, state3 }  => {
             (monero_wallet_restore_blockheight, transfer_proof, state3)
         }
@@ -39,7 +41,12 @@ pub async fn cancel(
 
         // Alice already in final state
         | AliceState::BtcRedeemed
-        | AliceState::XmrRefunded
+        | AliceState::XmrRefunded { .. }
+        | AliceState::BtcWithholdPublished { .. }
+        | AliceState::BtcWithholdConfirmed { .. }
+        | AliceState::BtcMercyGranted { .. }
+        | AliceState::BtcMercyPublished { .. }
+        | AliceState::BtcMercyConfirmed { .. }
         | AliceState::BtcEarlyRefundable { .. }
         | AliceState::BtcEarlyRefunded(_)
         | AliceState::BtcPunished { .. }